+++ audiolazy/lazy_misc.py
@@ -28,12 +28,14 @@ from math import pi
 import operator
 
 # Audiolazy internal imports
+from .internals import _deprecate
 from .lazy_compat import (xrange, xzip_longest, STR_TYPES, SOME_GEN_TYPES,
                           iteritems)
 from .lazy_core import StrategyDict
 
 __all__ = ["DEFAULT_SAMPLE_RATE", "rint", "blocks", "zero_pad", "elementwise",
-           "almost_eq", "sHz"]
+           "almost_eq", "sHz", "freq2lag", "lag2freq", "freq_to_lag",
+           "lag_to_freq"]
 
 DEFAULT_SAMPLE_RATE = 44100 # Hz (samples/second)
 
@@ -315,3 +317,15 @@ def sHz(rate):
 
   """
   return float(rate), 2 * pi / rate
+
+
+def freq2lag(v):
+  """ Converts from frequency (rad/sample) to lag (number of samples). """
+  return 2 * pi / v
+freq_to_lag = _deprecate(freq2lag)
+
+
+def lag2freq(v):
+  """ Converts from lag (number of samples) to frequency (rad/sample). """
+  return 2 * pi / v
+lag_to_freq = _deprecate(freq2lag)