+++ CHANGES.rst
@@ -32,6 +32,10 @@ AudioLazy changes history
 
   - Tox for testing with CPython 2.7, 3.2, 3.3 and 3.4, as well as PyPy
   - Continuous integration wih Travis CI and Coveralls
+  - Renamed ``freq2lag`` and ``lag2freq`` to use ``2`` instead of ``_to_``,
+    and moved to keep the unit converters on lazy_misc
+  - New ``internals.py`` module to avoid exposing package internals together
+    with the API
 
 + lazy_analysis: